+++ fhir_kindling/tests/test_fhir_server.py
@@ -1,5 +1,4 @@
 import os
-import pprint
 from unittest import mock
 
 import pytest
@@ -123,6 +122,13 @@ def test_server_rest_resources(fhir_server: FhirServer):
 
 def test_server_summary(fhir_server: FhirServer):
     summary = fhir_server.summary()
+    assert summary
+    print(summary)
+
+
+@pytest.mark.asyncio
+async def test_server_summary_async(fhir_server: FhirServer):
+    summary = await fhir_server.summary_async(display=False)
     assert summary
     print(summary)
 
@@ -558,18 +564,15 @@ def test_get_many(fhir_server: FhirServer):
 def test_transfer(fhir_server: FhirServer):
     api_url = os.getenv("FHIR_API_URL", "http://localhost:9090/fhir")
     server = FhirServer(api_address=api_url, timeout=None)
-    conditions = server.query("Condition").limit(10)
-    assert len(conditions.resources) == 10
-    transfer_url = os.getenv("TRANSFER_API_URL", "http://localhost:9091/fhir")
+    conditions = server.query("Condition")
+    transfer_url = os.getenv("TRANSFER_SERVER_URL", "http://localhost:9091/fhir")
     hapi_server = FhirServer(api_address=transfer_url, timeout=None)
 
     # conditions_before = hapi_server.query("Condition").all().resources
 
-    response = server.transfer(hapi_server, query_result=conditions)
+    response = server.transfer(hapi_server, query=conditions)
 
-    pprint.pp(response.create_responses)
-    assert len(response.create_responses) == 30
-    assert response.destination_server == hapi_server.api_address
+    assert response
 
 
 @pytest.mark.asyncio