+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.0.0] - 2023-
-Improve packaging with poetry to slim down library size with optional extras for datascience features.
-Add support for Python 3.9+. Code base cleanup and refactoring.
+Improve packaging with poetry to slim down library size with optional extras for datascience features. Record linkage
+for transfer functionality. Add support for Python 3.9+. Code base cleanup and refactoring.
 Big upgrades to the [documentation](https://migraf.github.io/fhir-kindling/) ðŸ‘€.
 
 
@@ -27,6 +27,9 @@ Big upgrades to the [documentation](https://migraf.github.io/fhir-kindling/) ðŸ‘
 ### Added
 - Optional progress bar for `summary`, `get_many`, `add_all` and `transfer` methods.
 - Additional property `resource_list` on `FhirQueryResponse` to get a list of all resources (even included ones) from the response.
+- `summary_async()` method for asynchronous server summary.
+- `total` property on `QueryResponse` to get the total number of resources returned by the query.
+- `record_linkage` functionality for `transfer()` method and split transfer functionality into separate module.
 
 ### Removed
 - Removed `to_dfs()` method on query response object. Use `flatten` function from the `serde` package instead.